Update zksync deposit script
