+++ scripts/config/local.json
@@ -1,6 +1,6 @@
 {
-  "zkSync": "0x82F67958A5474e40E1485742d648C0b0686b6e5D",
-  "wstETH": "0xb770Ea0F1762D73c8719B52eF981f7F1D824d9a7",
+  "zkSync": "0xabea9132b05a70803a4e85094fd0e1800777fbef",
+  "wstETH": "0x7f39c581f595b53c5cb19bd0b3f8da6c935e2ca0",
   "curve-stETH-pool": "0xDC24316b9AE028F1497c275EB9192a3Ea0f67022",
   "yearn-crvStETH-vault": "0xdCD90C7f6324cfa40d7169ef80b12031770B4325",
   "argent": {
